added "flasher"-subproject added testrom by axlr

+++ tools/crcbldr/trunk/CRCBLDR.c
@@ -6,8 +6,8 @@
 // #define thirdpart 1
 
 typedef unsigned char BYTE;
-typedef unsigned short int WORD;
-typedef unsigned int LONGWORD;
+typedef unsigned short WORD;
+typedef unsigned long LONGWORD;
 
 #define HEXLEN 0
 #define DATATYPE 1
